combine SignUpController.java with UserController.java
